WIP: Factor legend rendering to its own class
